add MemoryAgent class
