+++ core/cat/agents/procedures_agent.py
@@ -1,0 +1,7 @@
+
+
+
+class ProceduresAgent:
+
+    async def execute(self, stray):
+        pass
